Add support for unrecognized tags when querying images
Without a suitable .ExifTool_config file, custom tags are not listed in the
generated pstore tags.

This causes problems when trying to dump all tags embedded in the image:
unrecognized tags are returned as nil from the `self.original_tag` method and
thus are keyed on an empty string (and thus clobbered) in calls to, eg `to_hash`.

They will still be omitted from `all_tags` as this just returns all the
tags _known_ to Exiftool, ie contents of the pstore.

An additional instance-level `TagHash` is used to map unrecognized tags
and is combined in a new `original_tag` method which first attempts
retrieval from `MiniMagick.original_tag`, then the `@unmapped_tags`
TagHash.

For granular control of tags, use of an .ExifTool_config file is
required, which puts custom tags into the pstore and makes them
both read- and writable.